move packages + update setup.py + version bump to 0.3.0
