+++ deisa/__version__.py
@@ -9,4 +9,4 @@
 #
 ###################################################################################################
 
-__version__ = "0.2.5"
+__version__ = "0.3.0"
