--- setup.cfg
@@ -1,31 +1,0 @@
-# Copyright (c) 2020-2022 Centre national de la recherche scientifique (CNRS)
-# Copyright (c) 2020-2022 Commissariat a l'énergie atomique et aux énergies alternatives (CEA)
-# Copyright (c) 2020-2022 Institut national de recherche en informatique et en automatique (Inria)
-# Copyright (c) 2020-2022 Université Paris-Saclay
-# Copyright (c) 2020-2022 Université de Versailles Saint-Quentin-en-Yvelines
-#
-# SPDX-License-Identifier: MIT
-
-[metadata]
-name = deisa
-version = attr: deisa.__version__.__version__
-author = A. Gueroudji
-author_email = amal.gueroudji@cea.fr
-classifiers =
-    Programming Language :: Python :: 3.9
-    License :: OSI Approved :: MIT License
-    Operating System :: OS Independent
-    Development Status :: 3 - Alpha
-
-[options]
-package_dir=
-    =python-src
-packages=find:
-install_requires =
-    dask
-    dask[distributed]
-    PyYAML
-    numpy
-
-[options.packages.find]
-where=python-src
